fix: remove most mining code
Removed most mining code, as some is still needed for producing blocks or deep function calls in core (you would need to rewrite functions).

Will still try again to remove all, but this is pass 1 and we can test and see if antiviruses still pickup the program as a potential virus.

Fixes #25 from the community request.