fix: rd reverse address type check
Added reverse address type check for rd.

This is needed because if a user does not give out the .sxcash or .pxcash and instead gives out the public address the sender will not know if thats a special address or not. This safeguards against this as it will check the db for any regular address and return the type before sending, to correctly send regular address as well as namespace names.