feat: vote using any amount
Vote using any amount.
This creates a unstaked and staked balance atfer the vote. Sending tx will only use from the unstaked balance, unless you clear the staked outputs.

This will allow people to vote with certain amount and keep some in the wallet for transfers that will not break the reserve proof of the staked amount.